requirements.txt
dj-config-url==0.1.1
Django==2.2
feedparser==5.2.1
psycopg2==2.8.1
pytz==2018.9
sqlparse==0.3.0
django-tastypie==0.14.2
djangorestframework==3.9.2
django-filter==2.1.0